# teacher-ApiDocs.json
{"openapi":"3.0.1","info":{"title":"My API","description":"API description","version":"1.0.0"},"servers":[{"url":"http://localhost:8081","description":"Generated server url"}],"paths":{"/teachers/{teacher_id}/update-student":{"put":{"tags":["teacher-controller"],"operationId":"updateStudent","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Student"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Student"}}}}}}},"/teachers/update":{"put":{"tags":["teacher-controller"],"operationId":"updateTeacher","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Teacher"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Teacher"}}}}}}},"/teachers/insert":{"post":{"tags":["teacher-controller"],"operationId":"addTeacher","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Teacher"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Teacher"}}}}}}},"/teachers":{"get":{"tags":["teacher-controller"],"operationId":"getTeachers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Teacher"}}}}}}}},"/teachers/{teacher_id}":{"get":{"tags":["teacher-controller"],"operationId":"getTeacher","parameters":[{"name":"teacher_id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Teacher"}}}}}}},"/teachers/{teacher_id}/students":{"get":{"tags":["teacher-controller"],"operationId":"getStudentsByTeacherId","parameters":[{"name":"teacher_id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Student"}}}}}}}},"/teachers/{teacher_id}/delete-student/{rollno}":{"delete":{"tags":["teacher-controller"],"operationId":"deleteStudent","parameters":[{"name":"rollno","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/teachers/{rollno}":{"delete":{"tags":["teacher-controller"],"operationId":"deleteTeacher","parameters":[{"name":"teacher_id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"Student":{"type":"object","properties":{"rollno":{"type":"integer","format":"int32"},"name":{"type":"string"},"course":{"type":"string"},"address":{"type":"string"},"email":{"type":"string"},"age":{"type":"integer","format":"int32"},"gender":{"type":"string"},"parentName":{"type":"string"},"parentPhone":{"type":"string"},"admissionNumber":{"type":"string"},"admissionDate":{"type":"string","format":"date-time"},"teacher_id":{"type":"integer","format":"int32"}}},"Teacher":{"type":"object","properties":{"teacher_id":{"type":"integer","format":"int32"},"name":{"type":"string"},"subject":{"type":"string"},"phone":{"type":"string"},"address":{"type":"string"},"joinedOn":{"type":"string","format":"date-time"},"email":{"type":"string"},"qualification":{"type":"string"},"experience":{"type":"integer","format":"int32"},"salary":{"type":"number","format":"double"}}}}}}